[qubit pulse] fix
It is way more efficient (not only for performance but also for
understanding) to use a bitmap for the pulse parameter.

pulse can indicate the beginning, the end, none, or both.

A good way is to represent the beginning by 0b01, the end by 0b10, none
by 0, both by 0b01|0b10 = 0b11.