Update Build_on_windows.md
This is my first commit/PR to QUISP.
I was trying to install QUISP with OMNeT++ and faced some trouble importing eigen since I couldn't open OMNeT++ menu in property dialogue of quisp directory. I fixed this with the help of @Chibikuri, and now I'm adding the solution to windows build instruction.