[runtime]: add qubit_found flag for GET_QUBIT
previously, I implemented the GET_QUBIT can set error, but this is often
happend and it causes perfomance impact. and, the concept of error is
intended to describe something we cannot expect. if we can, it should be
a part of usual control flow. so I added the "qubit_found" flag and we
can handle the situation that qubit not found as simple branching.
therefore, I removed BNERR(branch if no error) instruction.