remove old es handling method from RuleEngine
the RuleEngine::updateResources_EntanglementSwapping method was
obsolated by the new RuleEngine::handleSwappingResult method. this
method uses the RuleEngine's Runtime and this change make the management
of the entangled qubit resources.
And also I need to cleanup the RuleEngine's unit tests cuz it was
totally messed up for no use of the google test fixture