add ComponentProvider::getNode
This change is required by the Router module. The Router module is used
in many types of components like BSANode. In this BSANode case, the
ComponentProvider cannot get the QNode because its BSANode because
the ComponentProvider has getQNode method, not getNode. This method
provides all types of Node includes QNode, BSANode, and SPDCNode in the
future